requirements.txt.txt
streamlit==1.35.0
opencv-python==4.9.0.80
numpy==1.26.4
joblib==1.4.2
Pillow==10.3.0
